refactor(RawFactEncoder): move to util module
